A Loan Tracker Demonstrating Basic SQLite Functions
An SQLite Database example that demonstrates adding, deleting, and
updating information in a database.  Three builds are included for major
operating systems (Windows, Mac, Linux).  The pre-build file is
available for viewing in Xojo (a free download) to view coding and SQL
calls.

Not all tabs are functional.  Here is a brief list of current
functionality

Filling out the fields and clicking "Add" adds data to the database

Highlighting a row in the listbox populates the text fields

Delete removes from the database, listbox, and clears the text fields

A highlighted field changes the "Add" button to "Update"
Changing a field and clicking "Update" changes the database, listbox,
clear the text fields, and changes the "Update" button to "Add"

If a field is highlighted and the "Select" button is pushed, the row
shows up on the listbox on the second panel

The entered data persists between application shutdowns and repopulates
upon opening the application